backend/requirements.txt
fastapi 
uvicorn 
groq 
python-dotenv 
requests 
pydantic 
python-multipart 
loguru
streamlit
pdfplumber
plotly